requirements.txt
flask==2.3.3
flask-cors==4.0.0
pymongo==4.5.0
python-dotenv==1.0.0
pyjwt==2.8.0
werkzeug==2.3.7
requests==2.31.0
python-slugify==8.0.1